feat(console): show Operation branch and folder on Cmd+K rows
- Extract the sidebar chip's workspace-context line into a shared
  OperationWorkspaceContext component with generic .operation-context styles
- Render the same branch/folder line inline after the name on Cmd+K
  Operation rows so each row stays one line; carry it in aria-label
- Add the branch changelog fragment
