+++ .changelog.d/palette-operation-context.md
@@ -1,0 +1,8 @@
+---
+branch: palette-operation-context
+---
+
+### fleet-console
+#### Changed
+- With "Show Operation location" on, the Cmd+K Operation rows now carry the same branch and working-folder line as the sidebar chips, inline after the name so each row stays one line.
+  ko: 「Operation 위치 표시」를 켜면 Cmd+K의 Operation 행에도 사이드바 칩과 같은 브랜치·작업 폴더 표시가 이름 옆 한 줄로 붙습니다.
